r-chess-frontend/requirements.txt
streamlit
python-chess
requests